+++ CHANGES.md
@@ -19,6 +19,7 @@
 * Add BTN api call parameter debug logging
 * Fix anime searches on BTN provider
 * Change replace "Daily-Search" with "Recent-Search"
+* Add daily search to recent search renaming to config migration code
 
 [develop changelog]